feat(rust): support static struct version hash check for rust (#2793)
## Why?

<!-- Describe the purpose of this PR. -->

## What does this PR do?

support struct version hash check for rust

## Related issues

<!--
Is there any related issue? If this PR closes them you say say
fix/closes:

- #xxxx0
- #xxxx1
- Fixes #xxxx2
-->

## Does this PR introduce any user-facing change?

<!--
If any user-facing interface changes, please [open an
issue](https://github.com/apache/fory/issues/new/choose) describing the
need to do so and update the document if necessary.

Delete section if not applicable.
-->

- [ ] Does this PR introduce any public API change?
- [ ] Does this PR introduce any binary protocol compatibility change?

## Benchmark

<!--
When the PR has an impact on performance (if you don't know whether the
PR will have an impact on performance, you can submit the PR first, and
if it will have impact on performance, the code reviewer will explain
it), be sure to attach a benchmark data here.

Delete section if not applicable.
-->

+++ java/fory-core/src/main/java/org/apache/fory/resolver/ClassInfo.java
@@ -58,7 +58,7 @@ public class ClassInfo {
       boolean isDynamicGeneratedClass,
       Serializer<?> serializer,
       short classId,
-      short xtypeId) {
+      int xtypeId) {
     this.cls = cls;
     this.fullNameBytes = fullNameBytes;
     this.namespaceBytes = namespaceBytes;
